restructured tools directory to organize tool usage and imports
