+++ backend/backend/__init__.py
@@ -1,15 +1,1 @@
 """Backend package initialization."""
-
-from backend.agent import ReactAgent, ToolRegistry
-from backend.tools import calculator, string_analyzer, list_processor, json_formatter, grade_python_assignment, run_pytest_on_directory
-
-__all__ = [
-    "ReactAgent",
-    "ToolRegistry",
-    "calculator",
-    "string_analyzer",
-    "list_processor",
-    "json_formatter",
-    "grade_python_assignment",
-    "run_pytest_on_directory"
-]
