+++ backend/backend/tools/__init__.py
@@ -1,19 +1,31 @@
 """Tools package initialization."""
 
+# Public tools
 from backend.tools.example_tools import (
     calculator,
     string_analyzer,
     list_processor,
     json_formatter,
+)
+
+from backend.tools.testing_tools import (
+    run_pytest_on_directory,
+)
+
+from backend.tools.grading_tools import (
     grade_python_assignment,
-    run_pytest_on_directory
 )
 
 __all__ = [
+    # Core example tools
     "calculator",
     "string_analyzer",
     "list_processor",
     "json_formatter",
+
+    # Grading-related tool
     "grade_python_assignment",
-    "run_pytest_on_directory"
+
+    # Test-running tool
+    "run_pytest_on_directory",
 ]
